chore: remove cfd account creation logic
