+++ packages/cfd/src/features/Containers/cfd-password-modal/cfd-password-modal.scss
@@ -1,0 +1,130 @@
+.cfd-password-manager {
+    margin: 5.6rem auto 3.2rem;
+    width: 100%;
+    height: 100%;
+
+    @media only screen and (max-height: 645px) {
+        margin-top: 1.5rem;
+        overflow: auto;
+    }
+
+    &--paragraph {
+        margin-bottom: 3.2rem;
+    }
+    &--error-message {
+        margin-bottom: 1em;
+    }
+    &__investor-wrapper {
+        margin: 3.2rem 0.5rem 0;
+
+        & .dc-input__label {
+            top: 0.9rem;
+        }
+        @include mobile {
+            padding-bottom: 12rem;
+        }
+
+        .cfd-password-manager--paragraph:first-child {
+            margin-bottom: 0.8rem;
+        }
+    }
+    .dc-password-meter__container,
+    .dc-password-input {
+        width: 30rem;
+        margin: auto;
+    }
+    &__investor-form {
+        width: 100%;
+        max-width: 300px;
+        margin: 2.4em auto 0;
+
+        @include mobile {
+            max-width: unset;
+        }
+    }
+    &__new-password {
+        margin-bottom: 8.6rem;
+    }
+    &__actions {
+        align-items: center;
+        display: flex;
+        flex-flow: column nowrap;
+    }
+    &--button {
+        margin-top: 1.6rem;
+
+        @include mobile {
+            width: 100%;
+            margin-top: 0.8rem;
+            max-width: 30rem;
+            white-space: normal;
+        }
+    }
+    &__success {
+        text-align: center;
+        margin-top: 3.2rem;
+
+        &-header {
+            @include typeface(--paragraph-center-bold-black);
+            margin: 1.6rem 0 0.8rem;
+        }
+        &-btn {
+            margin-top: 3.2rem;
+            width: 6.4rem;
+
+            @include mobile {
+                width: 100%;
+                max-width: 30rem;
+            }
+        }
+    }
+    .multi-step__header {
+        margin: 0 0 2.4rem;
+    }
+    .multi-step__component {
+        text-align: center;
+    }
+    @include desktop {
+        max-width: calc(450px + 1rem); // needs extra padding for the scrollbar, otherwise it will cover the words
+    }
+    @include mobile {
+        padding: 0 1.6rem;
+        margin-top: 2.4rem;
+
+        &__scroll-wrapper {
+            overflow-y: auto;
+            overflow-x: hidden;
+        }
+        /* iPhone SE screen width fixes due to UI space restrictions */
+        @media only screen and (max-width: 320px) {
+            padding: 0;
+        }
+    }
+}
+
+.dc-modal {
+    &__container {
+        min-width: initial;
+        &_cfd-password-modal {
+            &__description {
+                @include typeface(--xsmall-left-normal-black);
+                line-height: 1.5;
+                max-width: calc(min(100vw, 349px));
+                margin: -2.7rem 0 2.4rem;
+                color: var(--text-less-prominent);
+            }
+            &__account-title {
+                margin-bottom: 1.6rem;
+            }
+        }
+        &_cfd-dashboard__compare-accounts {
+            width: 904px;
+        }
+        &_cfd-password-manager__modal {
+            height: 100% !important;
+            min-height: calc(100vh - 22rem);
+            width: 904px;
+            max-height: calc(100vh - 48px - 36px) !important;
+        }
+    }
+}
