--- packages/cfd/src/features/Containers/cfd-password-modal/index.ts
@@ -1,3 +1,0 @@
-import CFDPasswordModal from './cfd-password-modal';
-
-export default CFDPasswordModal;
